requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.0.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-writable-nested==0.6.2
flake8==3.8.4
idna==2.9
importlib-metadata==1.6.0
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python3-openid==3.2.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
